[Refactor]#5 구리스가 enum 합치라고 해서 코드 수정했습니당
