+++ BOOKBBANG/BOOKBBANG.xcodeproj/project.xcworkspace/xcshareddata/swiftpm/Package.resolved
@@ -1,5 +1,5 @@
 {
-  "originHash" : "8646e25e180920a065232acf8b4ca4c4fdf4a57ffbb51e24fcf155613761b82d",
+  "originHash" : "f0ad6759a09c946f28101e6d15b40efaccb7d91e464f5f9a3d806f1e4d327c45",
   "pins" : [
     {
       "identity" : "waterfallgrid",
